게시판 등록 , 수정 , 삭제 , 임보자 게시판 등록 수정 삭제 mapper.xml , Controller , Service , Mapper 대폭 수정
xml이 잘못 작성되어있고 , 컨트롤러 , 서비스 , 매퍼도 잘못 작성되어 전체 수정을 진행 다만 재영님이 적용한 file 조건문은 주석 처리 후 진행함

+++ poinhanshin/src/main/java/com/project/poinhanshin/mapper/board/BoardMapper.java
@@ -27,7 +27,7 @@ public interface BoardMapper {
     int boardUpdateContent(BoardDto boardDto , Integer loginUser);
 
     // 게시물을 삭제한다.
-    int boardDeleteContent(Integer boardno , Long loginUser);
+    int boardDeleteContent(Integer boardno , Integer loginUser);
 
     // 최근 사용자가 등록한 게시물의 번호를 가져온다. (즐겨찾기?)
     int boardSelectRecentBoardNo(Integer board_userno);
